move out props from ModeContext
